Remove `mod helpers`, moving the ParseError to the crate root
